+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## [Unreleased]
 ### Added
-- `RateRequestAnimationComponent` to activate requests while an `AnimationClip` of the `Animation` component is playing.
+- `RateRequestAnimationComponent` to activate requests while `Animation` component is playing.
+- `RateRequestAnimatorComponent` to activate requests while `Animator` component is playing.
 ### Changed
 ### Deprecated
 ### Removed